refactor(bluetooth): Move BluetoothServerService to wearos package
Closes: https://github.com/CatimaLoyalty/Android/pull/3185#discussion_r3522868889
